made use of Nullable type strategy to read/write from/to buffer
